Handle Stdent Profile and Index Page

+++ AttendanceSystem/Models/Entities/Enrolllment.cs
@@ -4,7 +4,7 @@
     {
         public required string StudId { get; set; }
         public required int CrsId { get; set; }
-        public decimal? CrsAttendanceRate { get; set; }
+        public int? CrsAttendanceRate { get; set; }
         public  Student? student { get; set; }
         public  Course? course { get; set; }
     }
